+++ apps/bookit/src/routes/__layout-bookit.svelte
@@ -1,6 +1,0 @@
-<script>
-	import '@packages/components/global.css';
-  import '$bookit/app.css';
-</script>
-
-<slot />