+++ apps/bookit/svelte.config.js
@@ -1,7 +1,0 @@
-import { svelteConfig } from '@packages/config';
-
-const config = {
-  ...svelteConfig
-};
-
-export default config;