+++ apps/vitebook/.vitebook/App.svelte
@@ -1,7 +1,0 @@
-<script>
-  import '$components/global.css';
-
-  export let component;
-</script>
-
-<svelte:component this={component} />