+++ apps/vitebook/vite.config.js
@@ -1,8 +1,0 @@
-import { viteConfig } from '@packages/config';
-
-/** @type {import('vite').UserConfig} */
-const config = {
-  ...viteConfig
-};
-
-export default config;