+++ packages/components/lib/Scrollbar/Scrollbar.svelte
@@ -1,8 +1,0 @@
-<script>
-  import { Svrollbar } from 'svrollbar';
-  // import { Svrollbar, Svroller } from 'svrollbar';
-</script>
-
-<!-- <Svrollbar /> -->
-<!-- <Svrollbar initiallyVisible={true} /> -->
-<Svrollbar alwaysVisible={true} />