+++ packages/components/src/Scrollbar/Scrollbar.svelte
@@ -1,0 +1,9 @@
+<script>
+  import { Svrollbar } from 'svrollbar'
+  import '../Scrollbar/scrollbar.css'
+  // import { Svrollbar, Svroller } from 'svrollbar';
+</script>
+
+<!-- <Svrollbar /> -->
+<!-- <Svrollbar initiallyVisible={true} /> -->
+<Svrollbar alwaysVisible={true} />